+++ TabletopTweaks-Base/Bugfixes/Classes/Bloodrager.cs
@@ -116,7 +116,9 @@ namespace TabletopTweaks.Base.Bugfixes.Classes {
 
                     var BloodragerAbyssalBloodrageBonusProperty = BlueprintTools.GetBlueprint<BlueprintUnitProperty>("415f71e5a47f4cccb3dbd10bd7a0f8f8");
 
-                    BloodragerAbyssalBloodrageBonusProperty.BaseValue = 1;
+                    BloodragerAbyssalBloodrageBonusProperty.TemporaryContext(bp => { 
+                        bp.BaseValue = 1;
+                    });
 
                     TTTContext.Logger.LogPatch(BloodragerAbyssalBloodrageBonusProperty);
                 }